requirements.txt
accelerate
aiohttp
altair
bqplot
datasets
fastapi
filelock
geopandas
geopy
gitpython
gradio
huggingface-hub
ipywidgets
jupyterlab
matplotlib
numpy
opencv-python
pandas
pillow
plotly
protobuf
pyarrow
pyyaml
requests
scikit-learn
scipy
seaborn
spacy
sqlalchemy
streamlit
streamlit-folium
torch
torchaudio
torchvision
transformers
uvicorn
